refactor: rename project and update structure
