+++ tests/Unit/Common/GlobalUsings.cs
@@ -1,0 +1,4 @@
+global using Xunit;
+global using FluentAssertions;
+global using Moq;
+global using System.Threading.Tasks;
